requirements.txt
annotated-types==0.6.0
anyio==4.2.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.1
fastapi==0.109.0
h11==0.14.0
hiredis==2.3.2
idna==3.6
more-itertools==9.1.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
python-ulid==1.1.0
redis-om==0.2.1
sniffio==1.3.0
starlette==0.35.1
types-pyOpenSSL==24.0.0.20240130
types-redis==4.6.0.20240106
typing_extensions==4.9.0
uvicorn==0.27.0.post1
